analytics/requirements.txt
fastapi==0.101.1
pydantic==2.1.1
uvicorn==0.23.2
Faker==19.3.0
httpx==0.24.1
pytest == 7.4.0
pytest-cov == 4.1.0
requests == 2.31.0
snowflake-connector-python == 3.4.0
snowflake-snowpark-python == 1.10.0
Jinja2 == 3.1.2